+++ sdk/dotnet/Pulumi/Deployment/Deployment_Run.cs
@@ -3,7 +3,9 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
+using System.Linq;
 using System.Threading.Tasks;
+using Pulumi.Testing;
 
 namespace Pulumi
 {
@@ -87,20 +89,65 @@ namespace Pulumi
         public static Task<int> RunAsync<TStack>() where TStack : Stack, new()
             => CreateRunner().RunAsync<TStack>();
 
+        /// <summary>
+        /// Entry point to test a Pulumi application. Deployment will
+        /// instantiate a new stack instance based on the type passed as TStack
+        /// type parameter. This method creates no real resources.
+        /// Note: Currently, unit tests that call <see cref="TestAsync{TStack}"/>
+        /// must run serially; parallel execution is not supported.
+        /// </summary>
+        /// <param name="mocks">Hooks to mock the engine calls.</param>
+        /// <param name="options">Optional settings for the test run.</param>
+        /// <typeparam name="TStack">The type of the stack to test.</typeparam>
+        /// <returns>Test result containing created resources and errors, if any.</returns>
+        public static async Task<ImmutableArray<Resource>> TestAsync<TStack>(IMocks mocks, TestOptions? options = null) where TStack : Stack, new()
+        {
+            var engine = new MockEngine();
+            var monitor = new MockMonitor(mocks);
+            Deployment deployment;
+            lock (_instanceLock)
+            {
+                if (_instance != null)
+                    throw new NotSupportedException($"Mulitple executions of {nameof(TestAsync)} must run serially. Please configure your unit test suite to run tests one-by-one.");
+
+                deployment = new Deployment(engine, monitor, options);
+                Instance = deployment;
+            }
+
+            try
+            {
+                await deployment._runner.RunAsync<TStack>();
+                return engine.Errors.Count switch
+                {
+                    1 => throw new RunException(engine.Errors.Single()),
+                    int v when v > 1 => throw new AggregateException(engine.Errors.Select(e => new RunException(e))),
+                    _ => monitor.Resources.ToImmutableArray()
+                };
+            }
+            finally
+            {
+                lock (_instanceLock)
+                {
+                    _instance = null;
+                }
+            }
+        }
+
         private static IRunner CreateRunner()
         {
             // Serilog.Log.Logger = new LoggerConfiguration().MinimumLevel.Debug().WriteTo.Console().CreateLogger();
 
             Serilog.Log.Debug("Deployment.Run called.");
-            if (_instance != null)
+            lock (_instanceLock)
             {
-                throw new NotSupportedException("Deployment.Run can only be called a single time.");
-            }
+                if (_instance != null)
+                    throw new NotSupportedException("Deployment.Run can only be called a single time.");
 
-            Serilog.Log.Debug("Creating new Deployment.");
-            var deployment = new Deployment();
-            Instance = deployment;
-            return deployment._runner;
+                Serilog.Log.Debug("Creating new Deployment.");
+                var deployment = new Deployment();
+                Instance = deployment;
+                return deployment._runner;
+            }
         }
     }
 }