+++ CHANGELOG.md
@@ -13,12 +13,15 @@ CHANGELOG
 
 - Ensure Python overlays work as part of our SDK generation
   [#4043](https://github.com/pulumi/pulumi/pull/4043)
+
 - Fix terminal gets into a state where UP/DOWN don't work with prompts.
   [#4042](https://github.com/pulumi/pulumi/pull/4042)
 
 - Ensure old provider is not used when configuration has changed
   [#4051](https://github.com/pulumi/pulumi/pull/4051)
 
+- Support for unit testing and mocking in the .NET SDK.
+  [#3696](https://github.com/pulumi/pulumi/pull/3696)
 
 ## 1.12.0 (2020-03-04)
 - Avoid Configuring providers which are not used during preview.